feat[PRO-15] Ejecutar cierre contable #PRO-102

+++ src/pagalotodo-ucab-ms/UCABPagaloTodoMS.Application/Requests/DebtorsRequest.cs
@@ -3,6 +3,6 @@
 public class DebtorsRequest
 {
     public string? Identifier { get; set; }
-    public float? Amount { get; set; }
+    public double? Amount { get; set; }
     public bool? Status { get; set; }
 }
